requirements.txt
amqp==1.4.9
anyjson==0.3.3
attrs==16.2.0
billiard==3.3.0.23
bzip2==0.0.1
celery==3.1.20
cffi==1.8.3
cryptography==1.5.3
cssselect==1.0.0
decorator==4.0.10
Django==1.10.3
django-celery==3.1.17
django-dynamic-scraper==0.11.3
enum34==1.1.6
future==0.15.2
idna==2.1
ipaddress==1.0.17
jsonpath-rw==1.4.0
kombu==3.0.37
lxml==3.6.4
parsel==1.0.3
Pillow==3.4.2
ply==3.9
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
pysqlite==2.8.3
pytz==2016.7
queuelib==1.4.2
Scrapy==1.1.0
scrapy-djangoitem==1.1.0
service-identity==16.0.0
six==1.10.0
SQLAlchemy==1.1.3
Twisted==15.0.0
update==0.4.4
vine==1.1.3
w3lib==1.15.0
zope.interface==4.3.2
